api/requirements.txt
fastapi
uvicorn
python-telegram-bot==20.8
httpx